DDF-0000 dynamic fanout
